size_t/ptrdiff_t transition part 4: libast regex and string sublibraries
This is the fourth of the thickfold patch series, which enables ksh93
to operate within a 64-bit address space.

The parts of ksh93 affected by this commit are:
- The libast string library.
- The libast regex engine.

Change in the number of warnings on Linux when compiling with clang using
-Wsign-compare -Wshorten-64-to-32 -Wsign-conversion -Wimplicit-int-conversion:
3,214 => 2,617 => 248 (progression from part 3 => part 4 => part 13)

Progresses https://github.com/ksh93/ksh/issues/592

+++ src/lib/libast/regex/regerror.c
@@ -60,7 +60,7 @@ regerror(int code, const regex_t* p, char* buf, size_t size)
 	NoP(p);
 	if (code++ == REG_VERSIONID)
 		s = (const char*)fmtident(&id[1]);
-	else if (code >= 0 && code < elementsof(reg_error))
+	else if (code >= 0 && code < (ssize_t)elementsof(reg_error))
 		s = reg_error[code];
 	else
 		s = (const char*)"unknown error";
